+++ src/funding/funding.js
@@ -24,6 +24,7 @@ type IsFundingEligibleOptions = {|
   flow: $Values<typeof BUTTON_FLOW>,
   fundingEligibility: FundingEligibilityType,
   enableFunding?: $ReadOnlyArray<?$Values<typeof FUNDING>>,
+  disableFunding?: $ReadOnlyArray<?$Values<typeof FUNDING>>,
   components: $ReadOnlyArray<$Values<typeof COMPONENTS>>,
   onShippingChange: ?Function,
   onShippingAddressChange: ?Function,
@@ -32,9 +33,12 @@ type IsFundingEligibleOptions = {|
   wallet?: ?Wallet,
   applePaySupport: boolean,
   supportsPopups: boolean,
+  supportsVenmoPopups: boolean,
+  supportedNativeVenmoBrowser: boolean,
   supportedNativeBrowser: boolean,
   experiment?: Experiment,
   displayOnly?: $ReadOnlyArray<$Values<typeof DISPLAY_ONLY_VALUES>>,
+  userAgent?: string,
 |};
 
 function isFundingVaultable({
@@ -72,6 +76,7 @@ export function isFundingEligible(
     fundingSource,
     fundingEligibility,
     enableFunding,
+    disableFunding,
     components,
     onShippingChange,
     onShippingAddressChange,
@@ -81,11 +86,25 @@ export function isFundingEligible(
     wallet,
     applePaySupport,
     supportsPopups,
+    supportsVenmoPopups,
     supportedNativeBrowser,
+    supportedNativeVenmoBrowser,
     experiment,
     displayOnly,
+    userAgent,
   }: IsFundingEligibleOptions
 ): boolean {
+  // Temporary: Force credit to be eligible if the experiment is enabled
+  // but honor merchant opt-out via disableFunding
+  if (
+    source === FUNDING.CREDIT &&
+    experiment?.paypalCreditButtonCreateVaultSetupTokenExists &&
+    flow === BUTTON_FLOW.VAULT_WITHOUT_PURCHASE &&
+    !(disableFunding || []).includes(FUNDING.CREDIT)
+  ) {
+    return true;
+  }
+
   if (!fundingEligibility[source] || !fundingEligibility[source].eligible) {
     return false;
   }
@@ -156,10 +175,15 @@ export function isFundingEligible(
     return false;
   }
 
-  if (fundingConfig.requires) {
+  if (fundingConfig.requires && userAgent) {
     const required = fundingConfig.requires({ experiment, platform });
-
-    if (required.popup === true && supportsPopups === false) {
+    const popupSupport =
+      source === FUNDING.VENMO ? supportsVenmoPopups : supportsPopups;
+    const nativeBrowserSupport =
+      source === FUNDING.VENMO
+        ? supportedNativeVenmoBrowser
+        : supportedNativeBrowser;
+    if (required.popup === true && popupSupport === false) {
       return false;
     }
 
@@ -167,7 +191,7 @@ export function isFundingEligible(
       return false;
     }
 
-    if (required.native === true && supportedNativeBrowser === false) {
+    if (required.native === true && nativeBrowserSupport === false) {
       return false;
     }
   }
@@ -192,6 +216,7 @@ export function determineEligibleFunding({
   platform,
   fundingEligibility,
   enableFunding,
+  disableFunding,
   components,
   onShippingChange,
   onShippingAddressChange,
@@ -202,8 +227,11 @@ export function determineEligibleFunding({
   applePaySupport,
   supportsPopups,
   supportedNativeBrowser,
+  supportsVenmoPopups,
+  supportedNativeVenmoBrowser,
   experiment,
   displayOnly = [],
+  userAgent = "",
 }: {|
   fundingSource: ?$Values<typeof FUNDING>,
   remembered: $ReadOnlyArray<$Values<typeof FUNDING>>,
@@ -211,6 +239,7 @@ export function determineEligibleFunding({
   platform: $Values<typeof PLATFORM>,
   fundingEligibility: FundingEligibilityType,
   enableFunding?: $ReadOnlyArray<?$Values<typeof FUNDING>>,
+  disableFunding?: $ReadOnlyArray<?$Values<typeof FUNDING>>,
   components: $ReadOnlyArray<$Values<typeof COMPONENTS>>,
   onShippingChange?: ?Function,
   onShippingAddressChange?: ?Function,
@@ -221,8 +250,11 @@ export function determineEligibleFunding({
   applePaySupport: boolean,
   supportsPopups: boolean,
   supportedNativeBrowser: boolean,
+  supportsVenmoPopups: boolean,
+  supportedNativeVenmoBrowser: boolean,
   experiment: Experiment,
   displayOnly?: $ReadOnlyArray<$Values<typeof DISPLAY_ONLY_VALUES>>,
+  userAgent?: string,
 |}): $ReadOnlyArray<$Values<typeof FUNDING>> {
   if (fundingSource) {
     return [fundingSource];
@@ -235,6 +267,7 @@ export function determineEligibleFunding({
       fundingSource,
       fundingEligibility,
       enableFunding,
+      disableFunding,
       components,
       onShippingChange,
       onShippingAddressChange,
@@ -244,9 +277,12 @@ export function determineEligibleFunding({
       wallet,
       applePaySupport,
       supportsPopups,
+      supportsVenmoPopups,
       supportedNativeBrowser,
+      supportedNativeVenmoBrowser,
       experiment,
       displayOnly,
+      userAgent,
     })
   );