+++ src/marks/component.jsx
@@ -9,6 +9,7 @@ import {
   memoize,
   isApplePaySupported,
   supportsPopups as userAgentSupportsPopups,
+  getUserAgent,
 } from "@krakenjs/belter/src";
 import {
   PLATFORM,
@@ -18,6 +19,7 @@ import {
 import { getRememberedFunding } from "@paypal/funding-components/src";
 import {
   getEnableFunding,
+  getDisableFunding,
   getComponents,
   getFundingEligibility,
   getEnv,
@@ -28,14 +30,19 @@ import type {
   OnShippingAddressChange,
   OnShippingOptionsChange,
 } from "../ui/buttons/props";
-import { BUTTON_LAYOUT, BUTTON_FLOW } from "../constants";
+import { BUTTON_LAYOUT, BUTTON_FLOW, MARK_VARIATIONS } from "../constants";
 import { determineEligibleFunding, isFundingEligible } from "../funding";
 import {
+  supportsVenmoPopups,
+  isSupportedNativeVenmoBrowser,
+} from "../funding/util";
+import {
   isSupportedNativeBrowser,
-  getVenmoEligibility,
+  getButtonExperiments,
 } from "../zoid/buttons/util";
 
 import { MarksElement } from "./template";
+import { MarksElementRebrand } from "./templateRebrand";
 
 const DEFAULT_HEIGHT = 20;
 
@@ -46,10 +53,12 @@ type MarksInstance = {|
 
 type MarksProps = {|
   fundingSource?: ?$Values<typeof FUNDING>,
+  markVariation?: ?$Values<typeof MARK_VARIATIONS>,
   onShippingChange?: OnShippingChange,
   onShippingAddressChange?: OnShippingAddressChange,
   onShippingOptionsChange?: OnShippingOptionsChange,
   displayOnly?: $ReadOnlyArray<$Values<typeof DISPLAY_ONLY_VALUES>>,
+  userAgent: string,
 |};
 
 export type MarksComponent = (MarksProps) => MarksInstance;
@@ -57,10 +66,12 @@ export type MarksComponent = (MarksProps) => MarksInstance;
 export const getMarksComponent: () => MarksComponent = memoize(() => {
   function Marks({
     fundingSource,
+    markVariation,
     onShippingChange,
     onShippingAddressChange,
     onShippingOptionsChange,
     displayOnly,
+    userAgent = getUserAgent(),
   }: MarksProps = {}): MarksInstance {
     const height = DEFAULT_HEIGHT;
     const fundingEligibility = getFundingEligibility();
@@ -68,6 +79,7 @@ export const getMarksComponent: () => MarksComponent = memoize(() => {
     const remembered = getRememberedFunding();
     const layout = BUTTON_LAYOUT.VERTICAL;
     const enableFunding = getEnableFunding();
+    const disableFunding = getDisableFunding();
     const components = getComponents();
     const flow = BUTTON_FLOW.PURCHASE;
     const applePaySupport = fundingEligibility?.applepay?.eligible
@@ -75,7 +87,17 @@ export const getMarksComponent: () => MarksComponent = memoize(() => {
       : false;
     const supportsPopups = userAgentSupportsPopups();
     const supportedNativeBrowser = isSupportedNativeBrowser();
-    const experiment = getVenmoEligibility();
+    const experiment = getButtonExperiments();
+    const supportsVenmoPopup = supportsVenmoPopups(
+      experiment,
+      supportsPopups,
+      userAgent
+    );
+    const supportedNativeVenmoBrowser = isSupportedNativeVenmoBrowser(
+      experiment,
+      userAgent
+    );
+
     const hasShippingCallback = Boolean(
       onShippingChange || onShippingAddressChange || onShippingOptionsChange
     );
@@ -83,6 +105,7 @@ export const getMarksComponent: () => MarksComponent = memoize(() => {
       fundingSource,
       fundingEligibility,
       enableFunding,
+      disableFunding,
       components,
       platform,
       remembered,
@@ -95,8 +118,11 @@ export const getMarksComponent: () => MarksComponent = memoize(() => {
       hasShippingCallback,
       supportsPopups,
       supportedNativeBrowser,
+      supportsVenmoPopups: supportsVenmoPopup,
+      supportedNativeVenmoBrowser,
       experiment,
       displayOnly,
+      userAgent,
     });
     const env = getEnv();
 
@@ -111,6 +137,7 @@ export const getMarksComponent: () => MarksComponent = memoize(() => {
         fundingSource,
         fundingEligibility,
         enableFunding,
+        disableFunding,
         components,
         onShippingChange,
         onShippingAddressChange,
@@ -118,9 +145,12 @@ export const getMarksComponent: () => MarksComponent = memoize(() => {
         flow,
         applePaySupport,
         supportsPopups,
+        supportsVenmoPopups: supportsVenmoPopup,
         supportedNativeBrowser,
+        supportedNativeVenmoBrowser,
         experiment,
         displayOnly,
+        userAgent,
       });
     };
 
@@ -130,16 +160,29 @@ export const getMarksComponent: () => MarksComponent = memoize(() => {
           throw new Error(`${fundingSource || "marks"} not eligible`);
         }
 
+        const isRebrandEnabled = experiment?.isPaypalRebrandEnabled;
+
         getElement(container).appendChild(
           (
             <div>
-              <MarksElement
-                fundingEligibility={fundingEligibility}
-                fundingSources={fundingSources}
-                height={height}
-                experiment={experiment}
-                env={env}
-              />
+              {isRebrandEnabled ? (
+                <MarksElementRebrand
+                  fundingEligibility={fundingEligibility}
+                  fundingSources={fundingSources}
+                  markVariation={markVariation}
+                  height={height}
+                  experiment={experiment}
+                  env={env}
+                />
+              ) : (
+                <MarksElement
+                  fundingEligibility={fundingEligibility}
+                  fundingSources={fundingSources}
+                  height={height}
+                  experiment={experiment}
+                  env={env}
+                />
+              )}
             </div>
           ).render(dom({ doc: document }))
         );