+++ src/ui/buttons/buttons.jsx
@@ -143,6 +143,7 @@ export function Buttons(props: ButtonsProps): ElementNode {
     components,
     content,
     customerId,
+    disableFunding,
     displayOnly,
     enableFunding,
     env,
@@ -165,9 +166,12 @@ export function Buttons(props: ButtonsProps): ElementNode {
     style,
     supportedNativeBrowser,
     supportsPopups,
+    supportsVenmoPopups,
+    supportedNativeVenmoBrowser,
     userIDToken,
     vault,
     wallet,
+    userAgent,
   } = normalizeButtonProps(props);
   const { layout, shape, tagline } = style;
 
@@ -178,6 +182,7 @@ export function Buttons(props: ButtonsProps): ElementNode {
     platform,
     fundingEligibility,
     enableFunding,
+    disableFunding,
     components,
     onShippingChange,
     onShippingAddressChange,
@@ -188,8 +193,11 @@ export function Buttons(props: ButtonsProps): ElementNode {
     applePaySupport,
     supportsPopups,
     supportedNativeBrowser,
+    supportsVenmoPopups,
+    supportedNativeVenmoBrowser,
     experiment,
     displayOnly,
+    userAgent,
   });
   const multiple = fundingSources.length > 1;